debug

+++ pkg/microservice/user/core/repository/orm/role_binding.go
@@ -108,6 +108,10 @@ func DeleteRoleBindingByUID(uid, namespace string, db *gorm.DB) error {
 		return err
 	}
 
+	if len(resp) == 0 {
+		return nil
+	}
+
 	err = db.Delete(resp).Error
 
 	if err != nil {
